test: move GlobalFilter tests from Task
