+++ tests/Task.test.ts
@@ -1227,194 +1227,3 @@ describe('checking if task lists are identical', () => {
         expect(Task.tasksListsIdentical(list1, list2)).toBe(false);
     });
 });
-
-describe('check removal of the global filter', () => {
-    afterEach(() => {
-        GlobalFilter.reset();
-    });
-
-    type GlobalFilterRemovalExpectation = {
-        globalFilter: string;
-        markdownTask: string;
-        expectedDescription: string;
-    };
-
-    test.each<GlobalFilterRemovalExpectation>([
-        {
-            globalFilter: '#task',
-            markdownTask: '- [ ] #task this is a very simple task',
-            expectedDescription: 'this is a very simple task',
-        },
-        {
-            globalFilter: '',
-            markdownTask: '- [ ] #task this is a very simple task',
-            expectedDescription: '#task this is a very simple task',
-        },
-        {
-            globalFilter: '🞋',
-            markdownTask: '- [ ] task with emoji 🞋 global filter',
-            expectedDescription: 'task with emoji global filter',
-        },
-        {
-            globalFilter: '#t',
-            markdownTask: '- [ ] task with #t global filter in the middle',
-            expectedDescription: 'task with global filter in the middle',
-        },
-        {
-            globalFilter: '#t',
-            markdownTask: '- [ ] task with global filter in the end #t',
-            expectedDescription: 'task with global filter in the end',
-        },
-        {
-            globalFilter: '#t',
-            markdownTask: '- [ ] task with global filter in the end and some spaces  #t  ',
-            expectedDescription: 'task with global filter in the end and some spaces',
-        },
-        {
-            globalFilter: '#complex/global/filter',
-            markdownTask: '- [ ] task with #complex/global/filter in the middle',
-            expectedDescription: 'task with in the middle',
-        },
-        {
-            globalFilter: '#task',
-            markdownTask: '- [ ] task with an extension of the global filter #task/with/extension',
-            expectedDescription: 'task with an extension of the global filter #task/with/extension',
-        },
-        {
-            globalFilter: '#t',
-            markdownTask: '- [ ] task with #t multiple global filters #t',
-            expectedDescription: 'task with multiple global filters',
-        },
-        {
-            globalFilter: '#t',
-            markdownTask: '- [ ] #t', // confirm behaviour when the description is empty
-            expectedDescription: '',
-        },
-    ])(
-        'should parse "$markdownTask" and extract "$expectedDescription"',
-        ({ globalFilter, markdownTask, expectedDescription }) => {
-            // Arrange
-            GlobalFilter.set(globalFilter);
-
-            // Act
-            const task = constructTaskFromLine(markdownTask);
-
-            // Assert
-            expect(task).not.toBeNull();
-            expect(GlobalFilter.removeAsWordFrom(task!.description)).toEqual(expectedDescription);
-        },
-    );
-});
-
-describe('check removal of the global filter exhaustively', () => {
-    afterEach(() => {
-        GlobalFilter.reset();
-    });
-
-    type GlobalFilterRemoval = {
-        globalFilter: string;
-    };
-
-    function checkDescription(markdownLine: string, expectedDescription: string) {
-        const task = constructTaskFromLine(markdownLine);
-
-        // Assert
-        expect(task).not.toBeNull();
-        expect(GlobalFilter.removeAsWordFrom(task!.description)).toEqual(expectedDescription);
-    }
-
-    test.each<GlobalFilterRemoval>([
-        {
-            globalFilter: '#t',
-        },
-        // The characters listed below are the ones that are - or were - escaped by
-        // Task.escapeRegExp().
-        // See the developer.mozilla.org reference in that method.
-        // This test validates the escaping of each of those characters.
-        {
-            globalFilter: '.',
-        },
-        {
-            globalFilter: '*',
-        },
-        {
-            globalFilter: '+',
-        },
-        {
-            globalFilter: '?',
-        },
-        {
-            globalFilter: '^',
-        },
-        {
-            // Failed attempt at creating a failing test for when = was not escaped.
-            // When I make Task.escapeRegExp() escape =, I get:
-            // Invalid regular expression: /(^|\s)hello\=world($|\s)/: Invalid escape
-            globalFilter: 'hello=world',
-        },
-        {
-            // Failed attempt at creating a failing test for when ! was not escaped.
-            // When I make Task.escapeRegExp() escape !, I get:
-            // Invalid regular expression: /(^|\s)hello\!world($|\s)/: Invalid escape
-            globalFilter: 'hello!world',
-        },
-        {
-            // Failed attempt at creating a failing test for when : was not escaped.
-            // When I make Task.escapeRegExp() escape :, I get:
-            // Invalid regular expression: /(^|\s)hello\:world($|\s)/: Invalid escape
-            globalFilter: 'hello:world',
-        },
-        {
-            globalFilter: '$',
-        },
-        {
-            globalFilter: '{',
-        },
-        {
-            globalFilter: '}',
-        },
-        {
-            globalFilter: '(',
-        },
-        {
-            globalFilter: ')',
-        },
-        {
-            globalFilter: '|',
-        },
-        {
-            globalFilter: '[',
-        },
-        {
-            globalFilter: ']',
-        },
-        {
-            // Failed attempt at creating a failing test for when / was not escaped
-            globalFilter: '///',
-        },
-        {
-            globalFilter: '\\',
-        },
-    ])('should parse global filter "$globalFilter" edge cases correctly', ({ globalFilter }) => {
-        // Arrange
-        GlobalFilter.set(globalFilter);
-
-        // Act
-
-        // global filter removed at beginning, middle and end
-        let markdownLine = `- [ ] ${globalFilter} 1 ${globalFilter} 2 ${globalFilter}`;
-        let expectedDescription = '1 2';
-        checkDescription(markdownLine, expectedDescription);
-
-        // global filter not removed if non-empty non-tag characters before or after it
-        markdownLine = `- [ ] ${globalFilter}x 1 x${globalFilter} ${globalFilter}x 2 x${globalFilter}`;
-        expectedDescription = `${globalFilter}x 1 x${globalFilter} ${globalFilter}x 2 x${globalFilter}`;
-        checkDescription(markdownLine, expectedDescription);
-
-        // global filter not removed if non-empty sub-tag characters after it.
-        // Include at least one occurrence of global filter, so we don't pass by luck.
-        markdownLine = `- [ ] ${globalFilter}/x 1 x${globalFilter} ${globalFilter}/x 2 ${globalFilter} ${globalFilter}/x`;
-        expectedDescription = `${globalFilter}/x 1 x${globalFilter} ${globalFilter}/x 2 ${globalFilter}/x`;
-        checkDescription(markdownLine, expectedDescription);
-    });
-});
